requirements.txt
assemblyai
annotated-types          0.7.0
anyio                    4.4.0
assemblyai               0.27.0
asttokens                2.4.1
cachetools               5.3.3
certifi                  2024.6.2
charset-normalizer       3.3.2
colorama                 0.4.6
decorator                5.1.1
editdistance             0.8.1
exceptiongroup           1.2.1
executing                2.0.1
google-api-core          2.19.0
google-auth              2.30.0
google-cloud-core        2.4.1
google-cloud-speech      2.26.0
google-cloud-translate   3.15.3
googleapis-common-protos 1.63.1
grpcio                   1.64.1
grpcio-status            1.62.2
h11                      0.14.0
httpcore                 1.0.5
httpx                    0.27.0
idna                     3.7
ipython                  8.25.0
jedi                     0.19.1
joblib                   1.4.2
matplotlib-inline        0.1.7
numpy                    1.26.4
parso                    0.8.4
pip                      24.0
prompt_toolkit           3.0.46
proto-plus               1.23.0
protobuf                 4.25.3
pure-eval                0.2.2
pyasn1                   0.6.0
pyasn1_modules           0.4.0
PyAudio                  0.2.14
pydantic                 2.7.3
pydantic_core            2.18.4
Pygments                 2.18.0
requests                 2.32.3
rsa                      4.9
scikit-learn             1.5.0
scipy                    1.13.1
setuptools               58.1.0
six                      1.16.0
sniffio                  1.3.1
stack-data               0.6.3
threadpoolctl            3.5.0
traitlets                5.14.3
typing_extensions        4.12.2
urllib3                  2.2.1
wcwidth                  0.2.13
websockets               12.0